+++ public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="style.css" rel="stylesheet" />
+    <link href="./styles/style.css" rel="stylesheet" />
     <title>Document</title>
   </head>
   <body>
@@ -39,6 +39,15 @@
       </section>
       <footer></footer>
     </div>
+
+    <script type="module" src="../dist/elements.js"></script>
     <script type="module" src="../dist/index.js"></script>
+    <script type="module" src="../dist/numberBtnHandle.js"></script>
+    <script type="module" src="../dist/operatorBtnHandle.js"></script>
+    <script type="module" src="../dist/combineString.js"></script>
+    <script type="module" src="../dist/equlBtn.js"></script>
+    <script type="module" src="../dist/calculator.js"></script>
+    <script type="module" src="../dist/clear.js"></script>
+    <script type="module" src="../dist/render.js"></script>
   </body>
 </html>
