+++ tsconfig.json
@@ -4,7 +4,6 @@
     "module": "ESNext",
     "skipDefaultLibCheck": true,
     "outDir": "dist",
-    "moduleDetection": "force",
     "strict": true
   },
   "include": ["src"]
